+++ packages/core-js/modules/es.typed-array.sort.js
@@ -1,7 +1,10 @@
 'use strict';
+/* eslint-disable es/no-typed-arrays -- required for testing */
 var ArrayBufferViewCore = require('../internals/array-buffer-view-core');
+var global = require('../internals/global');
 var fails = require('../internals/fails');
-var $sort = require('../internals/array-sort');
+var aFunction = require('../internals/a-function');
+var arraySort = require('../internals/array-sort');
 var FF = require('../internals/engine-ff-version');
 var IE_OR_EDGE = require('../internals/engine-is-ie-or-edge');
 var V8 = require('../internals/engine-v8-version');
@@ -9,15 +12,23 @@ var WEBKIT = require('../internals/engine-webkit-version');
 
 var aTypedArray = ArrayBufferViewCore.aTypedArray;
 var exportTypedArrayMethod = ArrayBufferViewCore.exportTypedArrayMethod;
+var Uint16Array = global.Uint16Array;
+var nativeSort = Uint16Array && Uint16Array.prototype.sort;
 
-var STABLE_SORT = !fails(function () {
+// WebKit
+var ACCEPT_INCORRECT_ARGUMENTS = !!nativeSort && !fails(function () {
+  var array = new Uint16Array(2);
+  array.sort(null);
+  array.sort({});
+});
+
+var STABLE_SORT = !!nativeSort && !fails(function () {
   // feature detection can be too slow, so check engines versions
-  if (V8) return V8 < 73;
+  if (V8) return V8 < 74;
   if (FF) return FF < 67;
   if (IE_OR_EDGE) return true;
   if (WEBKIT) return WEBKIT < 602;
 
-  // eslint-disable-next-line es/no-typed-arrays -- required for testing
   var array = new Uint16Array(516);
   var expected = Array(516);
   var index, mod;
@@ -40,5 +51,7 @@ var STABLE_SORT = !fails(function () {
 // `%TypedArray%.prototype.sort` method
 // https://tc39.es/ecma262/#sec-%typedarray%.prototype.sort
 exportTypedArrayMethod('sort', function sort(comparefn) {
-  return $sort.call(aTypedArray(this), comparefn);
-}, !STABLE_SORT);
+  if (!STABLE_SORT) return arraySort.call(aTypedArray(this), comparefn);
+  if (comparefn !== undefined) aFunction(comparefn);
+  return nativeSort.call(this, comparefn);
+}, !STABLE_SORT || ACCEPT_INCORRECT_ARGUMENTS);